Introduce minor adjustments and fixes
Create a new WITH_FIPS_PROTOTYPE macro. This eases the 'decoration' of
exported functions, by defining the NSS FIPS prototype at the same time
we are defining the exported function.

In C_GetFunctionList, only decorate the functions if the call to
FC_GetFunctionList succeeded, and add a debug message when finished.

Also add a pFunctionList parameter to the DECORATE macro. This allows
us to call it in the context of other functions.