Implement exported function (symbols) decoration
Now DECLARE_DECORATOR() also works for symbols defined in a dependent
library (shared object). DECORATE_FUNCTION_LIST() is the way to replace
function pointers in a function list, saving the original function.

SAVE_ORIGINAL_SYMBOL() looks up for an overwritten exported symbol in
a dependent library, saving the original function for use in our own
library. This macro is called from the library constructor and uses
dlopen()/dlsym() to find the already loaded dependency, and resolve
the function we want to forward calls from the wrapper/decorator.

Use this mechanism to decorate FC_GetInterface and FC_GetFunctionList
from libsoftokn3.so, using these hook points to make the function list
pointers decorations. Implement C_GetInterface and C_GetFunctionList
forwarding to the FIPS versions (the ones we decorated).