Add test case
The test case must be run in a FIPS system, using a JDK which doesn't
have the Java (old) importer/exporter. Both conditions are checked by
the test before continuing the execution.

Usage:
  make test [JAVA=/path/to/java]

If no JAVA is provided, javac/java from the PATH is used.

This test is currently working and passing with the latest Temurin
builds of OpenJDK 8u, 11u, 17u and 21u.