Code review 1st iteration, part 2
In this review, besides minor changes, we improved the multithread
support of the debugging traces facility. Each debug message is now
synchronized with a mutex (avoiding interleaved message chunks), and we
added the thread-id (tid) as part of the message header.

In order to do so, the dbg_trace_attr() macro was moved from p11_util.h
to dbg_trace.h, and the should_dump_attr_value() function was moved from
p11_util.h to dbg_trace.c, renaming it with the "__dbg_" prefix.

In an aim to be able to build on older systems, we now link with
libpthread, although this isn't necessary in modern glibc versions:
https://developers.redhat.com/articles/2021/12/17/why-glibc-234-removed-libpthread

Co-authored-by: Martin Balao <mbalao@redhat.com>