Initialize NSS with NSS_NoDB_Init()
This way, we are sure we can safely use libnss3.so functionality without
breaking FIPS compliance. At the same time, we no longer need the
C_GetTokenInfo() hack to make NSS aware of the non-DB configuration.

Now that we call NSS_NoDB_Init() before C_Initialize(), the latter
returns CKR_CRYPTOKI_ALREADY_INITIALIZED, so we are tolerating this. We
decided not to remove the C_Initialize() call just in case some day NSS
decouples the software token initialization in a way that libnss3.so's
NSS_NoDB_Init() no longer implies libsoftokn3.so's C_Initialize().

NOTE: NSS_NoDB_Init() is equivalent to the OpenJDK setup when SunPKCS11
is configured passing 'nssDbMode = noDb', see:
https://github.com/openjdk/jdk/blob/jdk-21-ga/src/jdk.crypto.cryptoki/share/classes/sun/security/pkcs11/Secmod.java#L595-L604
https://github.com/openjdk/jdk/blob/jdk-21-ga/src/jdk.crypto.cryptoki/share/classes/sun/security/pkcs11/Config.java#L480-L489
https://github.com/openjdk/jdk/blob/jdk-21-ga/src/jdk.crypto.cryptoki/share/native/libj2pkcs11/j2secmod.c#L70-L71
https://github.com/openjdk/jdk/blob/jdk-21-ga/src/jdk.crypto.cryptoki/share/native/libj2pkcs11/j2secmod.c#L131-L136
https://github.com/nss-dev/nss/blob/NSS_3_97_RTM/lib/nss/nssinit.c#L885-L900
https://github.com/nss-dev/nss/blob/NSS_3_97_RTM/lib/nss/nssinit.c#L944-L953

The only difference is that NSS_NoDB_Init() additionally passes the
NSS_INIT_READONLY and NSS_INIT_OPTIMIZESPACE flags, but these shouldn't
make a great difference in absence of a database.