Code review 1st iteration, part 11: fix importer
This bug was uncovered after improving the test case to increase the
coverage of the PKCS11::getNativeKeyInfo() and PKCS11::createNativeKey()
code (from the JDK-6913047 native memory leak fix).

The test now executes two sign operations in a row with the imported
keys, to exercise the aforementioned JDK-6913047 code. With this test
change, we found that the importer code isn't undoing what the exporter
does regarding CKA_SENSITIVE (forcing it to CK_FALSE).

The importer now forces CKA_SENSITIVE=CK_TRUE when trying to import a
key, so that the FIPS NSS token's FC_UnwrapKey() does not reject it.

We found this because PKCS11::getNativeKeyInfo() received the key with
CKA_SENSITIVE=CK_FALSE from the exporter, and PKCS11::createNativeKey()
was trying to create (import) it with this very same attribute.

Co-authored-by: Martin Balao <mbalao@redhat.com>