Enable all providers in the test
Since Main::getInstance() takes care of getting the services from the
right provider (SunPKCS11 with libnssadapter.so, except when in data
generation mode), we can leave any other provider installed.

This reduces the pre- / post-Java 8 burden with remaining providers
installed for cross-provider checks. Also, since we are increasing
the providers list by 1, we don't need to insert empty providers to
terminate it.